remove stats for shapefile
