+++ urbanstats/website_data/table.py
@@ -1,6 +1,7 @@
 from functools import lru_cache
+
+from permacache import permacache, stable_hash
 from shapefiles import shapefiles_for_stats
-from stats_for_shapefile import compute_statistics_for_shapefile
 from urbanstats.data.census_blocks import RADII
 from urbanstats.data.census_histogram import census_histogram
 
@@ -14,7 +15,29 @@ from collections import Counter
 
 from urbanstats.data.gpw import compute_gpw_data_for_shapefile_table
 from urbanstats.special_cases.merge_international import merge_international_and_domestic
+from urbanstats.statistics.collections_list import statistic_collections
 from urbanstats.universe.annotate_universes import attach_intl_universes, attach_usa_universes
+
+
+@permacache(
+    "population_density/stats_for_shapefile/compute_statistics_for_shapefile_24",
+    key_function=dict(sf=lambda x: x.hash_key, statistic_collections=stable_hash),
+    multiprocess_safe=True,
+)
+def compute_statistics_for_shapefile(sf, statistic_collections=statistic_collections):
+    sf_fr = sf.load_file()
+    print(sf)
+    result = sf_fr[["shortname", "longname"]].copy()
+    result["area"] = sf_fr["geometry"].to_crs({"proj": "cea"}).area / 1e6
+    assert (result.longname == sf_fr.longname).all()
+    for k in sf.meta:
+        result[k] = sf.meta[k]
+
+    for collection in statistic_collections:
+        if collection.for_america():
+            collection.compute_statistics(sf, result, sf_fr)
+
+    return result
